Método para usuario

+++ controllers/usuarios.js
@@ -54,34 +54,76 @@ function obtenerUsuario(req, res) {
 }
 
 // Esta función modifica un usuario
-function modificarUsuario(req, res, next) {
-    const usr = Usuario.create({
+ function modificarUsuario(req, res, next) {
+    const usuario = Usuario.findByPk(req.params.id)
+    .then(user=>{
+        if(!user){
+            return res.status(404).json('El usuario no existe');
+        }
+        if(req.body.email){
+            user.email = req.body.email;
+        }
+        if(req.body.nombre){
+            user.nombre = req.body.nombre;
+        }
+        if(req.body.cargo){
+            user.cargo = req.body.cargo;
+        }
+        if(req.body.tipo){
+            user.tipo = req.body.tipo;
+        }
+        if(req.body.password){
+            user.crearPassword(req.body.password);
+        }
+        user.save().then(updatedUser => {
+            return res.status(201).json(updatedUser.toAuthJSON())
+        }).catch(next);
+    })
+    .catch(next);
+
+    /*
+    const usr = Usuario.build({
         id: req.params.id,
         ...req.body
     })
     usr.save().then(user => {
         return res.status(201).json(user.toAuthJSON())
     }).catch(next);
+    */
 }
 
 // Esta función elimina un usuario
 function eliminarUsuario(req, res) {
-    //const usr = Usuario.findByPk(req.params.id);
-    console.log(req.params.id);
     const usr = req.params.id;
-
     if (usr === null) {
         return res.sendStatus(401)
     } else {
-        /*usr.destroy().then(usr => {return res.status(200)
-        })
-        .catch(err => {
-            return res.sendStatus(500)
-        })
-    }*/
-        Usuario.destroy({ where: { idUsuario: usr } })
+        Usuario.findByPk(req.params.id).then(user=>{
+            if(!user){
+                return res.status(404).json('El usuario no existe');
+            }else{
+                Usuario.destroy({ where: { idUsuario: usr } })
+                .then(user => { res.status(200).json({ Operacion: 'Se ha borrado el registro con exito' }) })
+                .catch(error=>{
+                    return res.status(500).json('Error con el servidor');
+                })
+            }
+        }).catch(error=>{
+            
+        });
+        /*
+        const usuario =  Usuario.findByPk(req.params.id).catch()
+        if(!usuario){
+            return res.status(500).json('El usuario no existe');
+        }else{
+            console.log(usuario);
+            Usuario.destroy({ where: { idUsuario: usr } })
             .then(user => { res.status(200).json({ Operacion: 'Se ha borrado el registro con exito' }) })
-            .catch(res.sendStatus(500))
+            .catch(error=>{
+                return res.status(500).json('Error con el servidor');
+            })
+        }
+        */
     }
 }
 
@@ -116,47 +158,35 @@ function busquedaPorAtributos(req, res) {
     Usuario.findAll({
         attributes: ['idUsuario', 'email', 'nombre', 'cargo', 'departamento', 'tipo'],
         where: {
-
-
             [Op.or]: [{
                     email: {
-                        [Op.like]: `${atributoBusqueda}`
+                        [Op.like]: `%${atributoBusqueda}%`
                     }
                 },
-
                 {
                     nombre: {
-                        [Op.like]: `${atributoBusqueda}`
+                        [Op.like]: `%${atributoBusqueda}%`
                     }
                 },
                 {
                     departamento: {
-                        [Op.like]: `${atributoBusqueda}`
+                        [Op.like]: `%${atributoBusqueda}%`
                     }
                 },
                 {
                     cargo: {
-                        [Op.like]: `${atributoBusqueda}`
+                        [Op.like]: `%${atributoBusqueda}%`
                     }
                 },
                 {
                     tipo: {
-                        [Op.like]: `${atributoBusqueda}`
+                        [Op.like]: `%${atributoBusqueda}%`
                     }
                 }
             ]
-
-
-
-
-
-
-
-
-
         }
     }).then((users) => {
-        return /*res.json(users.map(u => u.map)*/ res.send(users);
+        return res.send(users);
 
     }).catch(error => {
         return res.status(401).send(error);
